Fixed the hung notes bug(I hope).
It appears that the official DSSI example plugin allow events to be dropped if the sample buffer runs out.

while (event_pos < event_count && pos >= events[event_pos].time.tick)

It sounds like a good idea in theory, but unfortunately those missed events are going to result in hung notes, or missed notes, which is worse than buffer underrun.

Commented out the && part of the if statement.