I had stated quite some time ago that I intended to dump Jack and go straight to ALSA...  This happened right around the time that horrible ALSA-epic-Xruns-on-USB-devices regression occurred, so I decided to wait until upstream reverted that commit before doing anything to the audio back-end.
Almost one year later, they still have yet to revert that commit, so in the
interest of not putting all of my eggs in one basket, instead of going
for ALSA I went for Portaudio.  By "not putting all of my eggs in one
basket", I mean that Windows and Mac ports of PyDAW will be evaluated soon,
which Portaudio makes possible.  People should at least have the option
of using an un-free OS for the sake of knowing that will have a actual
working audio stack.  Besides being cross platform, PyDAW now has a
device dialog like most Windows and Mac applications, and latency is
greatly improved.

If Ubuntu 14.04's ALSA still has the USB regression, I will consider not
supporting Linux as a tier-1 platform again until it is fixed, it's
getting hard to justify being Linux-only when the audio stack stays
broken for this long and all distros continue to package the latest version
anyways.  There really is no excuse for it, and only delusional
basement-dwellers would attempt to justify why musicians should consider
Linux a serious audio platform when it only works about 50% of the time.
For the n00bz that might be reading this, this is NOT the first time
this has happened to the audio stack, this is a recurring problem that
can only be fixed by having the discipline to do real regression testing
and actually revert the commits that break things.