Some fixes to audio item + removal of bar loop mode.
The rationale for removing bar mode is that there is no way to properly
loop through individual audio items in any app unless the item
boundaries align with the loop (otherwise it's impossible to
guarantee that it will sound good), and with the use of PyDAW as a
live performance tool on the roadmap, there is no place for features
that may make things sound bad.  People wanting to have a one bar
loop can just set a region length to one.